compiler: Refactor Absyn to add lexical context
Idents instead of being transparent and value based equality now are
opaque and do reference equality, which lets us compare bindings.

This is first in the series of patch to fix #7 and #63. This attemps to
adapt new absyn in compiler. Unit-tests and certain parts such as
`provide` are left out, but the compiler typechecks with current
changes.